grammar.py
This file makes use of the operations in units and adds a matching and a merging operation. First, a small symbolic grammar is converted to a vector form, then it is used to process a linguistic example "the green mouse". We perform match and then merge, showing how a small but complete example can be processed.